refactor(core): extract user update library
