+++ packages/core/src/routes/experience/classes/libraries/provision-library.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import { Component, CoreEvent, getEventName } from '@logto/app-insights/custom-event';
 import { appInsights } from '@logto/app-insights/node';
 import {
@@ -28,11 +27,7 @@ import type TenantContext from '#src/tenants/TenantContext.js';
 import { buildAppInsightsTelemetry } from '#src/utils/request.js';
 import { getTenantId } from '#src/utils/tenant.js';
 
-import {
-  type InteractionProfile,
-  type InteractionUserProvisioningProfile,
-  type WithHooksAndLogsContext,
-} from '../../types.js';
+import { type InteractionProfile, type WithHooksAndLogsContext } from '../../types.js';
 import { toUserSocialIdentityData } from '../utils.js';
 
 import {
@@ -54,15 +49,10 @@ type OrganizationProvisionPayload =
       organizationIds: string[];
     };
 
-type ProvisioningCustomDataOptions = {
-  mergeCustomData?: boolean;
-};
-
 type CreateUserOptions = {
   checkIdentifierCollision?: boolean;
-} & ProvisioningCustomDataOptions;
-
-type UpdateUserOptions = ProvisioningCustomDataOptions;
+  mergeCustomData?: boolean;
+};
 
 const mergeCreateUserCustomData = (
   existingCustomData: JsonObject | undefined,
@@ -80,17 +70,6 @@ const mergeCreateUserCustomData = (
     ? mergedCustomData
     : undefined;
 };
-
-const mergeUpdateUserCustomData = (
-  existingCustomData: JsonObject | undefined,
-  customData: JsonObject | undefined
-): JsonObject | undefined =>
-  customData && Object.keys(customData).length > 0
-    ? {
-        ...existingCustomData,
-        ...customData,
-      }
-    : undefined;
 
 export class ProvisionLibrary {
   constructor(
@@ -201,54 +180,6 @@ export class ProvisionLibrary {
     return user;
   }
 
-  async updateUser(
-    userId: string,
-    provisioningProfile: InteractionUserProvisioningProfile,
-    { mergeCustomData: shouldMergeCustomData = false }: UpdateUserOptions = {}
-  ) {
-    const { queries, libraries } = this.tenantContext;
-
-    await libraries.users.checkIdentifierCollision(
-      getProfileIdentifierCollisionPayload(provisioningProfile),
-      userId
-    );
-
-    const { passwordEncrypted, passwordEncryptionMethod, customData, profile, ...updateProfile } =
-      provisioningProfile;
-    const { customDataForUpdate, existingUser } = await this.resolveCustomDataForUpdate(
-      userId,
-      customData,
-      shouldMergeCustomData
-    );
-    const profileForUpdate = await this.resolveProfileForUpdate(
-      userId,
-      profile,
-      customDataForUpdate,
-      existingUser
-    );
-    const updatePayload = this.buildUpdateUserPayload({
-      updateProfile,
-      profileForUpdate,
-      customDataForUpdate,
-      passwordEncrypted,
-      passwordEncryptionMethod,
-    });
-
-    if (Object.keys(updatePayload).length === 0) {
-      return queries.users.findUserById(userId);
-    }
-
-    const user = await queries.users.updateUserById(
-      userId,
-      updatePayload,
-      customDataForUpdate === undefined ? 'merge' : 'replace'
-    );
-
-    this.ctx.appendDataHookContext('User.Data.Updated', { user });
-
-    return user;
-  }
-
   async addSsoIdentityToUser(
     userId: string,
     enterpriseSsoIdentity: Required<InteractionProfile>['enterpriseSsoIdentity']
@@ -291,85 +222,6 @@ export class ProvisionLibrary {
    * This method is used to get the provision context for a new user registration.
    * It will return the provision context based on the current tenant and the request context.
    */
-  private async resolveProfileForUpdate(
-    userId: string,
-    profile: InteractionUserProvisioningProfile['profile'],
-    customDataForUpdate: JsonObject | undefined,
-    existingUser?: User
-  ) {
-    const profilePatch =
-      profile !== undefined && Object.keys(profile).length > 0 ? profile : undefined;
-
-    if (profilePatch === undefined || customDataForUpdate === undefined) {
-      return profilePatch;
-    }
-
-    const user = existingUser ?? (await this.tenantContext.queries.users.findUserById(userId));
-
-    return {
-      ...user.profile,
-      ...profilePatch,
-    };
-  }
-
-  private buildUpdateUserPayload({
-    updateProfile,
-    profileForUpdate,
-    customDataForUpdate,
-    passwordEncrypted,
-    passwordEncryptionMethod,
-  }: {
-    updateProfile: Omit<
-      InteractionUserProvisioningProfile,
-      'passwordEncrypted' | 'passwordEncryptionMethod' | 'customData' | 'profile'
-    >;
-    profileForUpdate: InteractionUserProvisioningProfile['profile'];
-    customDataForUpdate: JsonObject | undefined;
-    passwordEncrypted: InteractionUserProvisioningProfile['passwordEncrypted'];
-    passwordEncryptionMethod: InteractionUserProvisioningProfile['passwordEncryptionMethod'];
-  }) {
-    return {
-      ...updateProfile,
-      ...conditional(profileForUpdate !== undefined && { profile: profileForUpdate }),
-      ...conditional(customDataForUpdate !== undefined && { customData: customDataForUpdate }),
-      ...conditional(
-        passwordEncrypted &&
-          passwordEncryptionMethod &&
-          buildUserPasswordPayload({
-            passwordEncrypted,
-            passwordEncryptionMethod,
-          })
-      ),
-    };
-  }
-
-  private async resolveCustomDataForUpdate(
-    userId: string,
-    customData: JsonObject | undefined,
-    shouldMergeCustomData: boolean
-  ): Promise<{ customDataForUpdate: JsonObject | undefined; existingUser?: User }> {
-    if (customData === undefined) {
-      return { customDataForUpdate: undefined };
-    }
-
-    if (!shouldMergeCustomData) {
-      return {
-        customDataForUpdate: Object.keys(customData).length > 0 ? customData : undefined,
-      };
-    }
-
-    if (Object.keys(customData).length === 0) {
-      return { customDataForUpdate: undefined };
-    }
-
-    const existingUser = await this.tenantContext.queries.users.findUserById(userId);
-
-    return {
-      customDataForUpdate: mergeUpdateUserCustomData(existingUser.customData, customData),
-      existingUser,
-    };
-  }
-
   private async getUserProvisionContext(profile: InteractionProfile): Promise<{
     /** Admin user provisioning flag */
     isCreatingFirstAdminUser: boolean;
@@ -530,4 +382,3 @@ export class ProvisionLibrary {
     });
   };
 }
-/* eslint-enable max-lines */
